Extract ScannerContext out of InputFileScannerContext
